switched back to read/show
