+++ example/Example/Concurrent.hs
@@ -18,12 +18,12 @@ page = do
 
 
 data Contents = Contents Milliseconds
-  deriving (Generic, ViewId)
+  deriving (Show, Read, ViewId)
 
 
 data ContentsAction
   = Load Int
-  deriving (Generic, ViewAction)
+  deriving (Show, Read, ViewAction)
 
 
 instance HyperView Contents where
