+++ example/Example/Redirects.hs
@@ -15,12 +15,12 @@ page = do
 
 
 data Contents = Contents
-  deriving (Generic, ViewId)
+  deriving (Show, Read, ViewId)
 
 
 data ContentsAction
   = RedirectAsAction
-  deriving (Generic, ViewAction)
+  deriving (Show, Read, ViewAction)
 
 
 instance HyperView Contents where
