+++ example/Example/Transitions.hs
@@ -15,13 +15,13 @@ page = do
 
 
 data Contents = Contents
-  deriving (Generic, ViewId)
+  deriving (Show, Read, ViewId)
 
 
 data ContentsAction
   = Expand
   | Collapse
-  deriving (Generic, ViewAction)
+  deriving (Show, Read, ViewAction)
 
 
 instance HyperView Contents where
