+++ src/Web/Hyperbole.hs
@@ -111,7 +111,6 @@ module Web.Hyperbole
   , view
   , ViewId
   , ViewAction
-  , Param
   , Response
 
     -- * Exports
@@ -144,7 +143,6 @@ import Web.Hyperbole.Effect
 import Web.Hyperbole.Embed
 import Web.Hyperbole.Forms (FormField, InputType (..), Validation (..), field, form, formField, input, invalidText, label, placeholder, submit, validate, validation)
 import Web.Hyperbole.HyperView
-import Web.Hyperbole.Param (Param)
 import Web.Hyperbole.Route
 import Web.Hyperbole.View
