+++ src/Web/Hyperbole/Param.hs
@@ -4,9 +4,12 @@ module Web.Hyperbole.Param where
 
 import Control.Applicative ((<|>))
 import Control.Monad (guard)
-import Data.Text (Text, pack)
+import Data.ByteString qualified as BS
+import Data.ByteString.Lazy qualified as BL
+import Data.Text (Text, pack, unpack)
 import Data.Text qualified as T
 import GHC.Generics
+import Text.Read (readMaybe)
 import Web.HttpApiData
 
 
@@ -132,3 +135,110 @@ instance Param String where
 parseParamQuery :: (FromHttpApiData a) => Text -> Maybe a
 parseParamQuery t =
   either (const Nothing) Just $ parseQueryParam t
+
+
+class GAction f where
+  gToAction :: f p -> Text
+  gParseAction :: Text -> Maybe (f p)
+
+
+instance (GAction f, GAction g) => GAction (f :*: g) where
+  gToAction (a :*: b) = gToAction a <> " " <> gToAction b
+
+
+  gParseAction t = do
+    let (at, bt) = breakProduct
+    a <- gParseAction at
+    b <- gParseAction bt
+    pure $ a :*: b
+   where
+    breakProduct
+      -- if it starts with a ( or a ", we want to skip until the terminator
+      | T.isPrefixOf "\"" t = breakString
+      | T.isPrefixOf "(" t = breakParens
+      | otherwise = breakSegment ' ' t
+
+    breakString =
+      let rest = T.drop 1 t
+       in ("\"" <> T.takeWhile (/= '"') rest <> "\"", T.drop 2 $ T.dropWhile (/= '"') rest)
+
+    breakParens =
+      let rest = T.drop 1 t
+       in ("(" <> T.takeWhile (/= ')') rest <> ")", T.drop 2 $ T.dropWhile (/= ')') rest)
+
+
+instance (GAction f, GAction g) => GAction (f :+: g) where
+  gToAction (L1 a) = gToAction a
+  gToAction (R1 b) = gToAction b
+  gParseAction t = do
+    (L1 <$> gParseAction @f t) <|> (R1 <$> gParseAction @g t)
+
+
+instance (Datatype d, GAction f) => GAction (M1 D d f) where
+  gToAction (M1 a) = gToAction a
+  gParseAction t = M1 <$> gParseAction t
+
+
+instance (Constructor c, GAction f) => GAction (M1 C c f) where
+  gToAction (M1 a) =
+    let cn = conName (undefined :: M1 C c f p)
+     in case gToAction a of
+          "" -> pack cn
+          t -> pack cn <> " " <> t
+  gParseAction t = do
+    let (c, rest) = breakSegment ' ' t
+    guard $ c == pack (conName (undefined :: M1 C c f p))
+    M1 <$> gParseAction rest
+
+
+instance GAction U1 where
+  gToAction _ = ""
+  gParseAction _ = pure U1
+
+
+instance (GAction f) => GAction (M1 S s f) where
+  gToAction (M1 a) = gToAction a
+  gParseAction t = M1 <$> gParseAction t
+
+
+instance GAction (K1 R Int) where
+  gToAction = toActionShow
+  gParseAction = parseActionRead
+
+
+instance GAction (K1 R Integer) where
+  gToAction = toActionShow
+  gParseAction = parseActionRead
+
+
+instance GAction (K1 R String) where
+  gToAction = toActionShow
+  gParseAction = parseActionRead
+
+
+instance GAction (K1 R Text) where
+  gToAction = toActionShow
+  gParseAction = parseActionRead
+
+
+instance GAction (K1 R BS.ByteString) where
+  gToAction = toActionShow
+  gParseAction = parseActionRead
+
+
+instance GAction (K1 R BL.ByteString) where
+  gToAction = toActionShow
+  gParseAction = parseActionRead
+
+
+instance {-# OVERLAPPABLE #-} (Show a, Read a) => GAction (K1 R a) where
+  gToAction (K1 a) = "(" <> pack (show a) <> ")"
+  gParseAction t = K1 <$> readMaybe (unpack t)
+
+
+toActionShow :: (Show a) => K1 R a p -> Text
+toActionShow (K1 a) = pack (show a)
+
+
+parseActionRead :: (Read a) => Text -> Maybe (K1 R a p)
+parseActionRead t = K1 <$> readMaybe (unpack t)
